quadrat mit runden ecken läuft jetzt, anpassungen in murmecha core waren nötig, 90 grad funktioniert

+++ src/runs_on_bot/quaternions.cpp
@@ -2,7 +2,7 @@
 
 // aufbau quaternionen
 struct Quaternion {
-    double w, x, y, z;
+    float w, x, y, z;
 
     // einheitsquaternion
     static Quaternion identity() {
